Abstract Class Repository
